Initializes Steam API and game server.
Initializes the Steam API at the start of the mod's execution and logs a fatal error if initialization fails.

Configures and logs into the Steam game server, advertising it to the master server.
Adds the ability to create a Steam lobby.

Changes `SetModDir` to take the `GameDescription` setting value instead of `GameDirectory`.